+++ mido/backends/test_backend.py
@@ -1,11 +1,5 @@
 from pytest import raises
-from .backend import Backend, _import_module
-
-def test_import_module():
-    _import_module('sys')
-    _import_module('os.path')
-    with raises(ImportError):
-        _import_module('!!!!!!!!!')
+from .backend import Backend
 
 def test_split_api():
     backend = Backend('test')